# apps/frontend/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/workspaces/{id}": {
      "get": {
        "operationId": "WorkspacesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get workspace detail",
        "tags": [
          "Workspaces"
        ]
      },
      "patch": {
        "operationId": "WorkspacesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update workspace",
        "tags": [
          "Workspaces"
        ]
      },
      "delete": {
        "operationId": "WorkspacesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete workspace",
        "tags": [
          "Workspaces"
        ]
      }
    },
...
    "/api/workspaces/{wsId}/comments": {
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a comment to an entity",
        "tags": [
          "Comments"
        ]
      },
      "get": {
        "operationId": "CommentsController_findByEntity",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List comments for an entity",
        "tags": [
          "Comments"
        ]
      }
    },
    "/api/workspaces/{wsId}/comments/{id}": {
      "patch": {
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit a comment (author only)",
        "tags": [
          "Comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a comment (author or admin)",
        "tags": [
          "Comments"
        ]
      }
    },
...
    "/api/workspaces/{wsId}/kb/comments/{commentId}": {
      "delete": {
        "operationId": "KnowledgeBaseController_deleteComment",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a comment",
        "tags": [
          "Knowledge Base"
        ]
      }
    },
...
    "/api/workspaces/{wsId}/dashboard/stats": {
      "get": {
        "operationId": "DashboardController_getStats",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dashboard aggregate stats: entity open/total counts, weekly created, monthly hours, recent activity",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/api/workspaces/{wsId}/dashboard": {
      "get": {
        "operationId": "DashboardsController_getDashboard",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "自定义仪表盘：读取工作区默认盘（无则 null）",
        "tags": [
          "dashboards"
        ]
      },
      "put": {
        "operationId": "DashboardsController_upsertDashboard",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertDashboardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "自定义仪表盘：保存默认盘布局（无则创建，有则更新）",
        "tags": [
          "dashboards"
        ]
      }
    },
    "/api/workspaces/{wsId}/dashboards": {
      "get": {
        "operationId": "DashboardsController_listDashboards",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "仪表盘列表（每工作区多盘）",
        "tags": [
          "dashboards"
        ]
      },
      "post": {
        "operationId": "DashboardsController_createDashboard",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertDashboardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "新建仪表盘",
        "tags": [
          "dashboards"
        ]
      }
    },
    "/api/workspaces/{wsId}/dashboards/{id}": {
      "put": {
        "operationId": "DashboardsController_updateDashboard",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertDashboardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "更新指定仪表盘（名称/布局）",
        "tags": [
          "dashboards"
        ]
      },
      "delete": {
        "operationId": "DashboardsController_deleteDashboard",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "删除指定仪表盘",
        "tags": [
          "dashboards"
        ]
      }
    },
    "/api/workspaces/{wsId}/report-subscriptions": {
      "get": {
        "operationId": "DashboardsController_listSubscriptions",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "定时报表订阅列表",
        "tags": [
          "dashboards"
        ]
      },
      "post": {
        "operationId": "DashboardsController_createSubscription",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "新建定时报表订阅（DAILY/WEEKLY，推送站内通知）",
        "tags": [
          "dashboards"
        ]
      }
    },
    "/api/workspaces/{wsId}/report-subscriptions/{id}": {
      "patch": {
        "operationId": "DashboardsController_updateSubscription",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "更新订阅（名称/频率/启停）",
        "tags": [
          "dashboards"
        ]
      },
      "delete": {
        "operationId": "DashboardsController_deleteSubscription",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "删除订阅（软删）",
        "tags": [
          "dashboards"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pinecone API",
    "description": "Aha!-style requirements management system.\n\nI10 API 治理（2026-08-18 P2）：全局前缀 /api；认证 = Bearer JWT（Web）或 API Token（pc_tok_ 前缀，SHA-256 存储）；全局限流 100 req/min/IP（Throttler）；Swagger 仅非生产暴露。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "pageId": {
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "description": "Parent comment ID for threaded replies"
          },
          "body": {
            "type": "string"
          }
        },
        "required": [
          "pageId",
          "body"
        ]
      },
...
      "UpsertDashboardDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSubscriptionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSubscriptionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}